requirements.txt
# Perception - AI News Intelligence Platform
# Python Dependencies

# Google Cloud ADK and AI Platform
google-genai[adk]>=0.6.0
google-cloud-aiplatform[adk,agent_engines]>=1.112.0
a2a-sdk>=0.3.4
google-cloud-firestore>=2.18.0
google-cloud-logging>=3.11.0
google-cloud-trace>=1.13.5

# Web server
uvicorn[standard]>=0.30.0

# Google Cloud services
google-cloud-discoveryengine>=1.5.0
cloudpickle>=3.0.0

# HTTP and async
httpx>=0.27.0
aiohttp>=3.10.0

# Data processing
pydantic>=2.9.0
python-dateutil>=2.9.0
pytz>=2024.1

# RSS and content parsing
feedparser>=6.0.11
beautifulsoup4>=4.12.3
lxml>=5.3.0

# Utilities
pyyaml>=6.0.2
python-dotenv>=1.0.1

# Observability
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-exporter-gcp-trace>=1.7.0
opentelemetry-instrumentation-httpx>=0.48b0

# Development
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
black>=24.8.0
flake8>=7.1.0
mypy>=1.11.0
